SpinDoctor

Documentation for SpinDoctor.

SpinDoctor.SpinDoctor

SpinDoctor.CosOGSE

SpinDoctor.DoublePGSE

SpinDoctor.PGSE

SpinDoctor.SinOGSE

SpinDoctor.TimeProfile

SpinDoctor.TimeProfile

SpinDoctor.assemble_flux_matrices

SpinDoctor.assemble_flux_matrix

SpinDoctor.assemble_mass_matrix

SpinDoctor.assemble_stiffness_matrix

SpinDoctor.bvalue_no_q

SpinDoctor.call_tetgen

SpinDoctor.compute_areas

SpinDoctor.compute_laplace_eig

SpinDoctor.compute_∇φ

SpinDoctor.constant_intervals

SpinDoctor.convexhull

SpinDoctor.couple_flux_matrix

SpinDoctor.create_cells

SpinDoctor.create_directions

SpinDoctor.create_directions

SpinDoctor.create_fibonacci_sphere

SpinDoctor.create_geometry

SpinDoctor.create_surfaces_cylinder

SpinDoctor.create_surfaces_neuron

SpinDoctor.create_surfaces_sphere

SpinDoctor.echotime

SpinDoctor.eig2length

SpinDoctor.fit_adc

SpinDoctor.fit_tensors

SpinDoctor.get_cmpt_volumes

SpinDoctor.get_coefficients

SpinDoctor.get_interval

SpinDoctor.get_mesh_surface

SpinDoctor.get_mesh_surfacenormals

SpinDoctor.get_mesh_volumes

SpinDoctor.gmesh2fem

SpinDoctor.integral

SpinDoctor.intervals

SpinDoctor.is_constant

SpinDoctor.length2eig

SpinDoctor.read_cells

SpinDoctor.read_surfaces

SpinDoctor.read_tetgen

SpinDoctor.save_cells

SpinDoctor.save_surfaces

SpinDoctor.save_tetgen

SpinDoctor.shapeder

SpinDoctor.solve_analytical

SpinDoctor.solve_btpde

SpinDoctor.solve_btpde_midpoint

SpinDoctor.solve_hadc

SpinDoctor.solve_karger

SpinDoctor.solve_mf

SpinDoctor.SpinDoctor — Module

Diffusion MRI modelling

SpinDoctor.CosOGSE — Type

f = CosOGSE(δ, Δ, nperiod)

Oscillating Gradient Spin Echo sequence with two cos-pulses of duration δ separated by a pause of duration Δ-δ for nperiod periods per pulse.

SpinDoctor.DoublePGSE — Type

f = DoublePGSE(δ, Δ)

Double Pulsed Gradient Spin Echo sequence with four pulses of duration δ, separated by pauses of duration Δ-δ, 0, and Δ-δ repsectively.

SpinDoctor.PGSE — Type

f = PGSE(δ, Δ)

Pulsed Gradient Spin Echo sequence with pulse duration δ and time between pulses Δ-δ.

SpinDoctor.SinOGSE — Type

f = SinOGSE(δ, Δ, nperiod)

Oscillating Gradient Spin Echo sequence with two sin-pulses of duration δ separated by a pause of duration Δ-δ for nperiod periods per pulse.

SpinDoctor.TimeProfile — Type

General time profile type (gradient sequence).

A TimeProfile should implement a calling method returning the value of the time profile at time t. It can optionally overwrite the echotime integral, and bvalue_no_q methods, the first providing a default value and the latter two computing numerical approximation to the integral quantities if not overwritten.

SpinDoctor.TimeProfile — Method

ft = (f::TimeProfile)(t)

Return the time profile value at time t.

SpinDoctor.assemble_flux_matrices — Method

assemble_flux_matrices(points, facets)

Assemble flux matrix (Q) for each compartment and boundary.

SpinDoctor.assemble_flux_matrix — Function

assemble_flux_matrix(facets, nodes[, weights])

Assemble 3D flux matrix using P1 finite elements.

This function is based on the Matlab function flux_matrixP1_3D.m from the Matlab nodal matrix assembly toolbox by Jan Valdman and Talal Rahman.

https://uk.mathworks.com/matlabcentral/fileexchange/27826-fast-fem-assembly-nodal-elements

Talal Rahman and Jan Valdman: Fast MATLAB assembly of FEM matrices in 2D and 3D: nodal elements, Applied Mathematics and Computation 219, 7151–7158 (2013).

SpinDoctor.assemble_mass_matrix — Function

assemble_mass_matrix(elements, volumes[, weights])

Assemble 3D mass matrix using P1 finite elements.

This function is based on the Matlab function mass_matrixP1_3D.m from the Matlab nodal matrix assembly toolbox by Jan Valdman and Talal Rahmnan:

https://uk.mathworks.com/matlabcentral/fileexchange/27826-fast-fem-assembly-nodal-elements

Talal Rahman and Jan Valdman: Fast MATLAB assembly of FEM matrices in 2D and 3D: nodal elements, Applied Mathematics and Computation 219, 7151–7158 (2013).

SpinDoctor.assemble_stiffness_matrix — Function

assemble_stiffness_matrix(elements, nodes[, D])

Assemble 3D stiffness matrix using P1 finite elements. A diffusion tensor D may be provided.

This function is based on the Matlab function stiffness_matrixP1_3D.m from the Matlab nodal matrix assembly toolbox by Jan Valdman and Talal Rahman.

https://uk.mathworks.com/matlabcentral/fileexchange/27826-fast-fem-assembly-nodal-elements

Talal Rahman and Jan Valdman: Fast MATLAB assembly of FEM matrices in 2D and 3D: nodal elements, Applied Mathematics and Computation 219, 7151–7158 (2013).

SpinDoctor.bvalue_no_q — Method

b = bvalue_no_q(f)

Compute the time profile contribution to the b-value. To obtain the b-value, multiply the result by q^2 = (γg)^2, where γ is the gyromagnetic ratio of the water proton, g is the gradient amplitude, and b = q^2 * bvalue_no_q(f).

SpinDoctor.call_tetgen — Function

call_tetgen(setup, surfaces)

Call Tetgen on surface geometry.

SpinDoctor.compute_areas — Method

compute_areas(facets, nodes)

Compute facet areas.

SpinDoctor.compute_laplace_eig — Function

compute_laplace_eig(mesh, pde, eiglim, neig_max)

Compute the Laplace eigenvalues, eigenfunctions and first order moments of products of pairs of eigenfunctions.

SpinDoctor.compute_∇φ — Method

compute_∇φ(nodes, points)

Compute the gradients of the finite element basis functions φ.

SpinDoctor.constant_intervals — Method

markers = constant_intervals(f)

Get markers for constant intervals. Default is false on all intervals.

SpinDoctor.convexhull — Method

elements, points = convexhull(points)

Compute the convex hull of a set of points points (of size dim * npoint). Return a matrix of boundary elements elements (of sizedim * nelement) and a restriction of the original points to the boundary (size dim * npoint_keep).

SpinDoctor.couple_flux_matrix — Function

couple_flux_matrix(model, Q_blocks[, symmetrical])

Create global flux matrix with coupling between compartments.

SpinDoctor.create_cells — Method

create_cells(setup)

Create geometrical configuration of cells. Return mathematical description (radii, centers).

SpinDoctor.create_directions — Method

create_directions(ndirection, flat)

Create ndirection directions unformly distributed in sphere. If flat is true, the directions lie in the plane instead of on the sphere.

SpinDoctor.create_directions — Method

create_directions(direction)

Get one normalized direction.

SpinDoctor.create_fibonacci_sphere — Method

create_fibonacci_sphere(npoint)

Create npoint evenly distributed points on the sphere.

SpinDoctor.create_geometry — Method

create_geometry(setup)

Create cells, surfaces and finite element mesh. This function does the following:

Check geometry setup consistency

Create or load cell configuration

Create or load surface triangulation

Call TetGen

Deform domain

Split mesh into compartments

For custom geometries with more than one compartment, call split_mesh directly instead. This requires facet and element labels.

SpinDoctor.create_surfaces_cylinder — Method

create_surfaces_cylinder(cells, setup)

Create surface triangulation of [inner and] outer cylinders [and ECS].

The ground surface is triangulated first, before the walls are "extruded" and the top surface is copied from the ground surface.

SpinDoctor.create_surfaces_neuron — Method

create_surfaces_neuron(filename, setup)

Create neuron surface mesh. A neuron surface mesh is loaded or create and loaded. An ECS can be added.

SpinDoctor.create_surfaces_sphere — Method

create_surfaces_sphere(cells, setup)

Create surface triangulation of [inner and] outer spheres [and ECS].

SpinDoctor.echotime — Method

TE = echotime(f)

Get echo time TE of the time profile f, which is the end of the last characteristic interval.

SpinDoctor.eig2length — Method

eig2length(λ, D)

Convert Laplace eigenvalue to length scale.

SpinDoctor.fit_adc — Method

fit_adc(b, S)

Fit Apparent Diffusion Coefficient (ADC) using a polynomial logfit of the normalized signal S against the b-values b.

SpinDoctor.fit_tensors — Method

fit_tensors(directions, adc)

Fit effective diffusion tensors to directionalized ADCs. The six components of the symmetric diffusion tensors are fitted by least squares to the gradient directions and resulting ADCs.

SpinDoctor.get_cmpt_volumes — Method

get_cmpt_volumes(mesh)

Get compartment volumes.

SpinDoctor.get_coefficients — Method

get_coefficients(setup)

Prepare PDE compartments.

SpinDoctor.get_interval — Method

get_interval(f, t)

Get the characteristic interval number of time profile f containing time t. If intervals[i] ≤ t < intervals[i+1], i is returned.

SpinDoctor.get_mesh_surface — Method

get_mesh_surface(points, facets)

Compute surface areas, centers and normals for each facet.

SpinDoctor.get_mesh_surfacenormals — Method

get_mesh_surfacenormals(points, elements, facets)

Computing outwards oriented surface normals on a connected mesh compartment.

SpinDoctor.get_mesh_volumes — Method

get_mesh_volumes(points, elements)

Get total volume and volumes of each tetrahedron of mesh.

SpinDoctor.gmesh2fem — Method

gmesh2fem(filename)

Extract points and elements from Gmesh file.

SpinDoctor.integral — Function

integral(f, t = echotime(f))

Integral of time profile f between 0 and t. Unless specified, the echotime is used as the upper integral limit.

For the PGSE, SinOGSE, CosOGSE and DoublePGSE sequences, analytical expressions are available. Otherwise a numerical integral is computed.

SpinDoctor.intervals — Method

intervals(f)

Get characteristic intervals of the time profile f.

SpinDoctor.is_constant — Method

is_constant(f, t)

Return true if time profile f is constant on the characteristic interval containing or starting at t.

SpinDoctor.length2eig — Method

length2eig(length, D)

Convert length scale to Laplace eigenvalue.

SpinDoctor.read_cells — Method

read_cells(cellfilename)

Read cell configuration (centers and radii).

SpinDoctor.read_surfaces — Method

read_surfaces(filename)

Read surface triangulation from file.

SpinDoctor.read_tetgen — Method

read_tetgen

Read mesh from Tetgen.

SpinDoctor.save_cells — Method

save_cells(cells, cellfilename)

Write geometrical cell configuration to file. The file contains the spatial dimension: 2 for cylinders, 3 for spheres.

SpinDoctor.save_surfaces — Method

save_surfaces(filename, surfaces)

Save surface triangulation. The files may be passed to Tetgen with filename.node.

SpinDoctor.save_tetgen — Method

save_tetgen(mesh_all, filename)

Save mesh in the Tetgen output format.

SpinDoctor.shapeder — Method

shapeder(points)

Derivative of shape functions with respect to reference coordinates (ξ₁, ξ₂, ξ₃).

SpinDoctor.solve_analytical — Function

solve_analytical(setup, experiment[, volumes])

Compute the signal in a multilayered cylinder or sphere using an analytical matrix formalism solution.

This function is based on the following articles and corresponding code: [1] D. S. Grebenkov, NMR Survey of Reflected Brownian Motion, Rev. Mod.Phys. 79, 1077 (2007) [2] D. S. Grebenkov, Pulsed-gradient spin-echo monitoring of restricted diffusion in multilayered structures, J. Magn. Reson. 205, 181-195 (2010).

SpinDoctor.solve_btpde — Method

solve_btpde(model, experiment)

Solve the Bloch-Torrey partial differential equation using P1 finite elements.

SpinDoctor.solve_btpde_midpoint — Method

solve_btpde_midpoint(model, setup)

Solve the Bloch-Torrey partial differential equation using P1 finite elements. This function uses a manual time stepping scheme (theta-rule), that requires a degree of implicitness θ and a time step dt. θ = 0.5: Crank-Nicolson (second order) θ = 1.0: Implicit Euler (first order)

SpinDoctor.solve_hadc — Method

solve_hadc(model, experiment) -> NamedTuple

Compute ADCs using the homogenized ADC model (HADC).

SpinDoctor.solve_karger — Method

solve_karger(model, experiment, difftensors)

Solve the finite pulse Karger model (FPK) using precomputed effective diffusion tensors difftensors.

SpinDoctor.solve_mf — Method

solve_mf(model, lap_eig, experiment)

Solve for magnetization using Matrix Formalism.