SpinDoctor

Documentation for SpinDoctor.

SpinDoctor.SpinDoctor

SpinDoctor.AbstractCallback

SpinDoctor.AbstractGradient

SpinDoctor.CosOGSE

SpinDoctor.DoublePGSE

SpinDoctor.GeneralGradient

SpinDoctor.Model

SpinDoctor.PGSE

SpinDoctor.Plotter

SpinDoctor.Printer

SpinDoctor.ScalarGradient

SpinDoctor.SinOGSE

SpinDoctor.TimeProfile

SpinDoctor.TimeProfile

SpinDoctor.VTKWriter

SpinDoctor.assemble_flux_matrices

SpinDoctor.assemble_flux_matrix

SpinDoctor.assemble_mass_matrix

SpinDoctor.assemble_matrices

SpinDoctor.assemble_stiffness_matrix

SpinDoctor.call_tetgen

SpinDoctor.coefficients

SpinDoctor.compute_areas

SpinDoctor.compute_mf_diffusion_tensor

SpinDoctor.compute_signal

SpinDoctor.compute_∇φ

SpinDoctor.convexhull

SpinDoctor.couple_flux_matrix

SpinDoctor.create_cells

SpinDoctor.create_fibonacci_sphere

SpinDoctor.create_geometry

SpinDoctor.create_surfaces

SpinDoctor.create_surfaces

SpinDoctor.create_surfaces

SpinDoctor.echotime

SpinDoctor.echotime

SpinDoctor.eig2length

SpinDoctor.fit_adc

SpinDoctor.fit_tensors

SpinDoctor.get_cmpt_volumes

SpinDoctor.get_mesh_surface

SpinDoctor.get_mesh_surfacenormals

SpinDoctor.get_mesh_volumes

SpinDoctor.gmesh2fem

SpinDoctor.int_F²

SpinDoctor.integral

SpinDoctor.intervals

SpinDoctor.isconstant

SpinDoctor.length2eig

SpinDoctor.limit_lengthscale

SpinDoctor.output_type

SpinDoctor.plot_field

SpinDoctor.plot_mesh

SpinDoctor.radial_dimension

SpinDoctor.read_cells

SpinDoctor.read_surfaces

SpinDoctor.read_tetgen

SpinDoctor.save_cells

SpinDoctor.save_surfaces

SpinDoctor.save_tetgen

SpinDoctor.shapeder

SpinDoctor.solve

SpinDoctor.solve

SpinDoctor.solve

SpinDoctor.solve

SpinDoctor.solve

SpinDoctor.solve

SpinDoctor.solve

SpinDoctor.solve

SpinDoctor.solve_multigrad

SpinDoctor.split_field

SpinDoctor.unitcircle

SpinDoctor.unitsphere

SpinDoctor.SpinDoctor — Module

Diffusion MRI modelling

SpinDoctor.AbstractCallback — Type

AbstractCallback

Callback to be performed after every time step when solving the BTPDE. The callback is initialized before time stepping using the initialize! function, updated after every step using update!, and finalized after time stepping using finalize!.

SpinDoctor.AbstractGradient — Type

AbstractGradient{T}

Magnetic field gradient $\vec{g}(t)$.

SpinDoctor.CosOGSE — Type

f = CosOGSE(δ, Δ, nperiod)

Oscillating Gradient Spin Echo sequence with two cos-pulses of duration δ separated by a pause of duration Δ-δ for nperiod periods per pulse.

SpinDoctor.DoublePGSE — Type

f = DoublePGSE(δ, Δ)

Double Pulsed Gradient Spin Echo sequence with four pulses of duration δ, separated by pauses of duration Δ-δ, 0, and Δ-δ repsectively.

SpinDoctor.GeneralGradient — Type

GeneralGradient(g⃗)

General gradient sequence g⃗(t) ∈ R³. The direction and amplitude may vary in time.

SpinDoctor.Model — Type

Model(mesh, ρ, D, T₂, κ, γ)

Discretized biological model with initial spin densities ρ, diffusion tensors D, T₂-relaxation times T₂, wall permeabilities κ, and gyromacnetic ratio γ (defaults to the one of water protons).

SpinDoctor.PGSE — Type

f = PGSE(δ, Δ)

Pulsed Gradient Spin Echo sequence with pulse duration δ and time between pulses Δ-δ.

SpinDoctor.Plotter — Type

Plotter(; nupdate = 1)

Plot the evolution of the BTPDE during time stepping. This requires loading the GLMakie plotting backend (]add GLMakie; using GLMakie). The plot is updated every nupdate time step. The resulting figure contains a plot of the time profile, total signal attenuation, and magnetization field (complex magnitude and phase shift).

SpinDoctor.Printer — Type

Printer(; nupdate = 1)

Print time stepping information to the console.

SpinDoctor.ScalarGradient — Type

ScalarGradient(dir, profile, amplitude)

Gradient sequence with amplitude amplitude, normal direction dir and scalar time profile profile.

The direction is constant, while the amplitude is controlled by the time profile.

The direction is normalized upon construction.

SpinDoctor.SinOGSE — Type

f = SinOGSE(δ, Δ, nperiod)

Oscillating Gradient Spin Echo sequence with two sin-pulses of duration δ separated by a pause of duration Δ-δ for nperiod periods per pulse.

SpinDoctor.TimeProfile — Type

TimeProfile

General time profile type (gradient sequence).

A TimeProfile should implement a calling method returning the value of the time profile at time t. It can optionally overwrite the echotime integral, and int_F² methods, the first providing a default value and the latter two computing numerical approximation to the integral quantities if not overwritten.

SpinDoctor.TimeProfile — Method

ft = (f::TimeProfile)(t)

Return the time profile value at time t.

SpinDoctor.VTKWriter — Type

VTKWriter(; nupdate = 1, dir = "output", filename = "solution")

Write magnetization field to a VTK file after every nupdate timestep. The files are stored in a ParaView data collection file (PVD). The magnetization time series may be visualized in ParaView by opening the file "$dir/$filename.pvd". The compartments are labeled.

SpinDoctor.assemble_flux_matrices — Method

assemble_flux_matrices(points, facets)

Assemble flux matrix (Q) for each compartment and boundary.

SpinDoctor.assemble_flux_matrix — Function

assemble_flux_matrix(facets, nodes[, weights])

Assemble 3D flux matrix using P1 finite elements.

This function is based on the Matlab function flux_matrixP1_3D.m from the Matlab nodal matrix assembly toolbox by Jan Valdman and Talal Rahman.

https://uk.mathworks.com/matlabcentral/fileexchange/27826-fast-fem-assembly-nodal-elements

Talal Rahman and Jan Valdman: Fast MATLAB assembly of FEM matrices in 2D and 3D: nodal elements, Applied Mathematics and Computation 219, 7151–7158 (2013).

SpinDoctor.assemble_mass_matrix — Function

assemble_mass_matrix(elements, volumes[, weights])

Assemble 3D mass matrix using P1 finite elements.

This function is based on the Matlab function mass_matrixP1_3D.m from the Matlab nodal matrix assembly toolbox by Jan Valdman and Talal Rahmnan:

https://uk.mathworks.com/matlabcentral/fileexchange/27826-fast-fem-assembly-nodal-elements

Talal Rahman and Jan Valdman: Fast MATLAB assembly of FEM matrices in 2D and 3D: nodal elements, Applied Mathematics and Computation 219, 7151–7158 (2013).

SpinDoctor.assemble_matrices — Method

assemble_matrices(model)

Assemble finite element matrices.

SpinDoctor.assemble_stiffness_matrix — Function

assemble_stiffness_matrix(elements, nodes[, D])

Assemble 3D stiffness matrix using P1 finite elements. A diffusion tensor D may be provided.

This function is based on the Matlab function stiffness_matrixP1_3D.m from the Matlab nodal matrix assembly toolbox by Jan Valdman and Talal Rahman.

https://uk.mathworks.com/matlabcentral/fileexchange/27826-fast-fem-assembly-nodal-elements

Talal Rahman and Jan Valdman: Fast MATLAB assembly of FEM matrices in 2D and 3D: nodal elements, Applied Mathematics and Computation 219, 7151–7158 (2013).

SpinDoctor.call_tetgen — Function

call_tetgen(setup, surfaces)

Call Tetgen on surface geometry.

SpinDoctor.coefficients — Method

coefficients(setup; D, T₂, ρ, κ, γ)

Prepare PDE compartments.

SpinDoctor.compute_areas — Method

compute_areas(facets, nodes)

Compute facet areas.

SpinDoctor.compute_mf_diffusion_tensor — Function

compute_mf_diffusion_tensor

Compute an effective diffusion tensors using the matrix formalism.

SpinDoctor.compute_signal — Method

compute_signal(M, ξ) = sum(M * ξ)

Compute signal from magnetization ξ, using the mass matrix M for integration.

Given a mesh mesh and a vector of compartment mass matrices M_cmpts, a vector of compartment signals may be obtained by compute_signal.(M_cmpts, split_field(mesh, ξ)).

SpinDoctor.compute_∇φ — Method

compute_∇φ(nodes, points)

Compute the gradients of the finite element basis functions φ.

SpinDoctor.convexhull — Method

elements, points = convexhull(points)

Compute the convex hull of a set of points points (of size dim * npoint). Return a matrix of boundary elements elements (of sizedim * nelement) and a restriction of the original points to the boundary (size dim * npoint_keep).

SpinDoctor.couple_flux_matrix — Function

couple_flux_matrix(model, Q_blocks[, symmetrical])

Create global flux matrix with coupling between compartments.

SpinDoctor.create_cells — Method

create_cells(setup)

Create geometrical configuration of cells. Return mathematical description (radii, centers).

SpinDoctor.create_fibonacci_sphere — Method

create_fibonacci_sphere(npoint)

Create npoint evenly distributed points on the sphere.

SpinDoctor.create_geometry — Method

create_geometry(setup)

Create cells, surfaces and finite element mesh. This function does the following:

Check geometry setup consistency

Create or load cell configuration

Create or load surface triangulation

Call TetGen

Deform domain

Split mesh into compartments

For custom geometries with more than one compartment, call split_mesh directly instead. This requires facet and element labels.

SpinDoctor.create_surfaces — Method

create_surfaces(setup::CylinderSetup, cells)

Create surface triangulation of [inner and] outer cylinders [and ECS].

The ground surface is triangulated first, before the walls are "extruded" and the top surface is copied from the ground surface.

SpinDoctor.create_surfaces — Method

create_surfaces(setup::NeuronSetup, filename)

Create neuron surface mesh. A neuron surface mesh is loaded or create and loaded. An ECS can be added.

SpinDoctor.create_surfaces — Method

create_surfaces(setup::SphereSetup, cells)

Create surface triangulation of [inner and] outer spheres [and ECS].

SpinDoctor.echotime — Method

echotime(gradient)

Get echo time TE of gradient.

SpinDoctor.echotime — Method

echotime(f)

Get echo time TE of the time profile f, which is the end of the last characteristic interval.

SpinDoctor.eig2length — Method

eig2length(λ, D)

Convert Laplace eigenvalue to length scale.

SpinDoctor.fit_adc — Method

fit_adc(bvalues, signals)

Fit the apparent diffusion coefficient (ADC) using a polynomial logfit of the normalized signals signals against the b-values bvalues.

SpinDoctor.fit_tensors — Method

fit_tensors(directions, adcs)

Fit effective diffusion tensors to directionalized ADCs. The six components of the symmetric diffusion tensors are fitted by least squares to the gradient directions and resulting ADCs.

SpinDoctor.get_cmpt_volumes — Method

get_cmpt_volumes(mesh)

Get compartment volumes.

SpinDoctor.get_mesh_surface — Method

get_mesh_surface(points, facets)

Compute surface areas, centers and normals for each facet.

SpinDoctor.get_mesh_surfacenormals — Method

get_mesh_surfacenormals(points, elements, facets)

Computing outwards oriented surface normals on a connected mesh compartment.

SpinDoctor.get_mesh_volumes — Method

get_mesh_volumes(points, elements)

Get total volume and volumes of each tetrahedron of mesh.

SpinDoctor.gmesh2fem — Method

gmesh2fem(filename)

Extract points and elements from Gmesh file.

SpinDoctor.int_F² — Method

b = int_F²(f)

Compute the time profile contribution to the b-value. To obtain the b-value, multiply the result by q^2 = (γg)^2, where γ is the gyromagnetic ratio of the water proton, g is the gradient amplitude, and b = q^2 * int_F²(f).

SpinDoctor.integral — Function

integral(f, t = echotime(f))

Integral of time profile f between 0 and t. Unless specified, the echotime is used as the upper integral limit.

For the PGSE, SinOGSE, CosOGSE and DoublePGSE sequences, analytical expressions are available. Otherwise a numerical integral is computed.

SpinDoctor.intervals — Function

intervals(f)

Get characteristic intervals of the time profile f.

SpinDoctor.isconstant — Method

isconstant(profile)

Return true if the time profile profile is intervalwise constant.

SpinDoctor.length2eig — Method

length2eig(length, D)

Convert length scale to Laplace eigenvalue.

SpinDoctor.limit_lengthscale — Method

limit_lengthscale(lap_eig, λ_max)

Only keep modes with length scales larger than minimum

SpinDoctor.output_type — Function

output_type(problem)

Get type returned by solve(problem).

SpinDoctor.plot_field — Method

plot_field(femesh, ξ)

Plot field ξ on the finite element mesh.

SpinDoctor.plot_mesh — Method

plot_mesh(femesh)

Plot finite element mesh.

SpinDoctor.radial_dimension — Function

radial_dimension(setup)

Get dimension of radial direction.

SpinDoctor.read_cells — Method

read_cells(cellfilename)

Read cell configuration (centers and radii).

SpinDoctor.read_surfaces — Method

read_surfaces(filename)

Read surface triangulation from file.

SpinDoctor.read_tetgen — Method

read_tetgen

Read mesh from Tetgen.

SpinDoctor.save_cells — Method

save_cells(cells, cellfilename)

Write geometrical cell configuration to file. The file contains the spatial dimension: 2 for cylinders, 3 for spheres.

SpinDoctor.save_surfaces — Method

save_surfaces(filename, surfaces)

Save surface triangulation. The files may be passed to Tetgen with filename.node.

SpinDoctor.save_tetgen — Method

save_tetgen(mesh_all, filename)

Save mesh in the Tetgen output format.

SpinDoctor.shapeder — Method

shapeder(points)

Derivative of shape functions with respect to reference coordinates (ξ₁, ξ₂, ξ₃).

SpinDoctor.solve — Function

solve

Solve problem.

SpinDoctor.solve — Method

solve(problem::AnalyticalMatrixFormalism, gradient, volumes = nothing)

Compute the signal in a multilayered cylinder or sphere using an analytical matrix formalism solution.

This function is based on the following articles and corresponding code: [1] D. S. Grebenkov, NMR Survey of Reflected Brownian Motion, Rev. Mod.Phys. 79, 1077 (2007) [2] D. S. Grebenkov, Pulsed-gradient spin-echo monitoring of restricted diffusion in multilayered structures, J. Magn. Reson. 205, 181-195 (2010).

SpinDoctor.solve — Method

solve_hadc(problem::HADC, gradient::ScalarGradient)

Compute the ADC using a homogenized ADC model (HADC). This is currently only implemented for scalar gradients.

SpinDoctor.solve — Method

solve(problem::Karger, gradient)

Solve the finite pulse Karger model (FPK) using precomputed effective diffusion tensors difftensors.

SpinDoctor.solve — Method

solve(laplace::Laplace)

Compute the Laplace eigenvalues, eigenfunctions and first order moments of products of pairs of eigenfunctions.

SpinDoctor.solve — Method

solve(problem::MatrixFormalism, matrices, lap_eig, gradient)

Solve for magnetization using Matrix Formalism.

SpinDoctor.solve — Method

solve(problem::GeneralBTPDE, gradient[; callbacks])

Solve the Bloch-Torrey partial differential equation using P1 finite elements.

SpinDoctor.solve — Method

run(problem::IntervalConstanBTPDE, gradient)

Solve the Bloch-Torrey partial differential equation using P1 finite elements. This function uses a manual time stepping scheme (theta-rule), that requires a degree of implicitness θ and a time step Δt. θ = 0.5: Crank-Nicolson (second order) θ = 1.0: Implicit Euler (first order)

SpinDoctor.solve_multigrad — Method

solve_multigrad(problem, gradients)

Solve problem for multiple magnetic field gradients.

SpinDoctor.split_field — Method

split_field(mesh, ξ)

Split the field ξ into a vector containing a view of each compartment.

SpinDoctor.unitcircle — Method

unitcircle(ndirection; half = false, normal = [0, 0, 1])

Create ndirection directions unformly distributed on the unit circle defined by normal.

If half is true, the directions lie on a half-circle instead of the whole circle.

SpinDoctor.unitsphere — Method

unitsphere(ndirection; half = false, normal = [0, 0, 1])

Create ndirection directions unformly distributed on the unit sphere.

If half is true, the points will be discributed on the hemisphere defined by normal.